Improve Equality Checks for Red Tree
Elements in the red tree are equal if they wrap the same green node at
the same position. This means we don't _need_ to memoise red children
to ensure that equality checks work.

Another advantage here is we can drop down to simple reference equality
for the wrapped nodes. The underlying green node will always return the
same children on each enumeration so we can save on structual equality
checks for larger nodes.